fix: TTS float16 output, chunked ASR dtype mismatch, doctor GPU compat
- TTS: convert Bark output from float16 to float32 before scipy.io.wavfile
- Chunked ASR: cast input_features to model dtype (float16) before generate
- Doctor CLI: use getattr for total_memory/total_mem PyTorch compat

All pipelines tested on KazNU A10 GPUs: text→text, audio→text, image→text,
text→image, image→image, video→text, text→speech, document→text,
document→document, code, streaming, chunked ASR, encoder router.

+++ python/omniff/models/tts.py
@@ -52,7 +52,7 @@ class TTSModel(OmniModel):
                 input_ids[k] = v.to(self._model.device)
 
         speech_output = self._model.generate(**input_ids)
-        audio_array = speech_output[0].cpu().numpy().squeeze()
+        audio_array = speech_output[0].cpu().float().numpy().squeeze()
 
         sample_rate = self._model.generation_config.sample_rate
